fix: resolve dependency cycle in world types
Moves CharacterArchetype to a dedicated type definition file to prevent
circular dependency between types and implementation utilities.
This resolves the dependency-cruiser violation causing CI failures.
